src/main/java/edu/jyu/sell/service/impl/ProductInfoServiceImpl.java
package edu.jyu.sell.service.impl;

import edu.jyu.sell.entity.ProductInfo;
import edu.jyu.sell.repository.ProductInfoRepository;
import edu.jyu.sell.service.ProductInfoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 商品详情业务类
 *
 * @author Jason
 * @create 2017-12-04 21:26
 **/
@Service
public class ProductInfoServiceImpl implements ProductInfoService {

    @Autowired
    private ProductInfoRepository repository;

    @Override
    public ProductInfo findOne(String productId) {
        return repository.findOne(productId);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return repository.findByProductStatus(ProductInfo.PRODUCT_STATUS_UP);
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return repository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return repository.save(productInfo);
    }
}
